requirements.txt
certifi==2019.11.28
chardet==3.0.4
ffmpeg-python==0.2.0
future==0.18.2
idna==2.8
oauthlib==3.1.0
pkg-resources==0.0.0
PySocks==1.7.1
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
tweepy==3.8.0
urllib3==1.25.8
